+++ src/util.h
@@ -11,11 +11,13 @@
 
 long GetTimeMS();
 
-INLINE float Sigmoid(float s) { return 1.0f / (1.0f + expf(-s * SS)); }
+INLINE float Sigmoid(float s) { return 1.0f / (1.0f + expf(-s * (float)SS)); }
 
-INLINE float SigmoidPrime(float s) { return s * (1.0 - s) * SS; }
+INLINE float SigmoidPrime(float s) { return s * (1.0f - s) * (float)SS; }
 
-INLINE float ReLUPrime(float s) { return s > 0; }
+INLINE float ReLUPrime(float s) { return s > 0.0f; }
+
+//INLINE float CReLUPrime(float s) { return s > 0.0f && s < (float)CRELU_MAX; } // (0, 256)
 
 INLINE uint64_t NetworkHash(NN* nn) {
   uint64_t hash = 0;
